pyproject.toml
[project]
name = "market-analyzer"
version = "0.1.0"
description = ""
authors = [
    {name = "jcnaef",email = "jcnaef@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "firebase-admin (>=6.0.0,<7.0.0)",
    "thefuzz (>=0.22.1,<0.23.0)",
    "groq (>=1.1.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "market_analyzer", from = "src"}]

[tool.pytest.ini_options]
testpaths = ["tests"]
spec_header_format = "{module_path}:"
spec_test_format = "{result} {name}"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-spec (>=5.2.0,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
